put_db_data/dummy_data_status/requirements.txt
psycopg2-binary
python-dotenv
sqlalchemy
tqdm